pyproject.toml
[tool.poetry]
name = "credit-card-transaction-airflow-etl-project"
version = "0.1.0"
description = "This ETL (Extract, Transform, Load) project focuses on processing and analyzing a dataset of credit card transactions. The main objective is to convert raw data into valuable and actionable information for business decision-making."
authors = ["Sebastian Ortiz, David Cajiao Lazt, Juan Andrés López"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
numpy = "^2.1.1"
pandas = "^2.2.3"
psycopg2 = "^2.9.9"
requests = "^2.32.3"
scipy = "^1.14.1"
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
